Refactor configuration and logging settings for improved readability and consistency
- Adjusted formatting in `defaults.go` for module log defaults to enhance clarity.
- Cleaned up whitespace in `env.go`, `examples_test.go`, and `security_helpers_test.go` for better code style.
- Updated test cases in `species_tracking_test.go` and `tls.go` to maintain consistent formatting.
- Changed logging parameters from `int` to `uint64` in `mqtt/client.go` to ensure proper data type usage.
- Enhanced comments in `birdnet.go` and `datastore.go` for better understanding of metrics recording.
- Improved code structure in `notification.go` and `myaudio.go` by aligning comments and variable declarations.
- Conducted whitespace cleanup in various test files to adhere to coding standards.
- Ensured all functions in `securefs.go` and `fuzz_test.go` maintain consistent error handling and logging practices.

+++ internal/datastore/resource_monitor.go
@@ -337,7 +337,7 @@ func ValidateStartupDiskSpace(dbPath string) error {
 	}
 
 	log.Info("Startup disk space validation passed",
-		logger.Int64("available_mb", int64(availableMB)),
+		logger.Uint64("available_mb", availableMB),
 		logger.Int64("required_mb", MinDiskSpaceStartup),
 		logger.String("path", dir))
